move all schema related files to schema folder merge migration manger in schema manager
